requirements.txt
opencv-python
pillow
numpy
tqdm

mss

pynput